Generate ZKR here instead of in zirgen
