+++ risc0/circuit/rv32im-m3-sys/cxx/WORKSPACE
@@ -136,7 +136,7 @@ maybe(
 
 git_repository(
     name = "zirgen",
-    commit = "c942819362ad1edbfa2da241d1055ef2b4cf79aa",
+    commit = "c942819362ad1edbfa2da241d1055ef2b4cf79aa",  # TODO: update commit
     remote = "https://github.com/risc0/zirgen.git",
 )
