+++ risc0/circuit/rv32im-m3-sys/cxx/examples/hello/hello.h
@@ -1,0 +1,25 @@
+// Copyright 2025 RISC Zero, Inc.
+//
+// Licensed under the Apache License, Version 2.0, <LICENSE-APACHE or
+// http://apache.org/licenses/LICENSE-2.0> or the MIT license <LICENSE-MIT or
+// http://opensource.org/licenses/MIT>, at your option. This file may not be
+// copied, modified, or distributed except according to those terms.
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+//
+// SPDX-License-Identifier: Apache-2.0 OR MIT
+
+#include "prove/prove.h"
+#include "verify/verify.h"
+
+namespace hello_m3 {
+
+risc0::CircuitInfo* getCircuitInfo(risc0::IHalPtr hal, size_t po2);
+
+risc0::VerifyCircuitInfo* getVerifyCircuitInfo();
+
+} // end namespace hello_m3
